Add tests for resource selectors
