+++ cmd/kubernetes-admission-controller/main_test.go
@@ -203,166 +203,6 @@ func TestConfig(t *testing.T) {
 
 }
 
-func TestDoesObjectMatchResourceSelector(t *testing.T) {
-	testCases := []struct {
-		name      string
-		selector  ResourceSelector
-		assertion func(t *testing.T, actual bool)
-	}{
-		{
-			name:      "should match anything",
-			selector:  ResourceSelector{PodSelectorType, ".*", ".*"},
-			assertion: assertShouldMatch,
-		},
-		{
-			name:      "should match key 'label.*'",
-			selector:  ResourceSelector{PodSelectorType, "label.*", ".*"},
-			assertion: assertShouldMatch,
-		},
-		{
-			name:      "should NOT match key '^label$'",
-			selector:  ResourceSelector{PodSelectorType, "^label$", ".*"},
-			assertion: assertShouldNotMatch,
-		},
-		{
-			name:      "should match key 'label'",
-			selector:  ResourceSelector{PodSelectorType, "label", ".*"},
-			assertion: assertShouldMatch,
-		},
-		{
-			name:      "should match key 'labelowner'",
-			selector:  ResourceSelector{PodSelectorType, "labelowner", ".*"},
-			assertion: assertShouldMatch,
-		},
-		{
-			name:      "should match key 'labelowner' with value 'lsometeam'",
-			selector:  ResourceSelector{PodSelectorType, "labelowner", "lsometeam"},
-			assertion: assertShouldMatch,
-		},
-		{
-			name:      "should match key 'labelowner' with value 'lsome'",
-			selector:  ResourceSelector{PodSelectorType, "labelowner", "lsome"},
-			assertion: assertShouldMatch,
-		},
-		{
-			name:      "should match key 'annotation.*'",
-			selector:  ResourceSelector{PodSelectorType, "annotation.*", ".*"},
-			assertion: assertShouldMatch,
-		},
-		{
-			name:      "should match key 'annotationowner' with value 'asometeam'",
-			selector:  ResourceSelector{PodSelectorType, "annotationowner", "asometeam"},
-			assertion: assertShouldMatch,
-		},
-		{
-			name:      "should match key 'own' with value '.*team'",
-			selector:  ResourceSelector{PodSelectorType, "own", ".*team"},
-			assertion: assertShouldMatch,
-		},
-		{
-			name:      "should NOT match key 'notfound'",
-			selector:  ResourceSelector{PodSelectorType, "notfound", ".*"},
-			assertion: assertShouldNotMatch,
-		},
-		{
-			name:      "should NOT match value 'anotherteam'",
-			selector:  ResourceSelector{PodSelectorType, ".*", "anotherteam"},
-			assertion: assertShouldNotMatch,
-		},
-		{
-			name:      "should NOT match key 'owner' with value 'anotherteam'",
-			selector:  ResourceSelector{PodSelectorType, "owner", "anotherteam"},
-			assertion: assertShouldNotMatch,
-		},
-	}
-
-	metadata := testMetadata()
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			actual := doesObjectMatchResourceSelector(&metadata, testCase.selector)
-			testCase.assertion(t, actual)
-		})
-	}
-}
-
-func TestMatchImageResource(t *testing.T) {
-	testCases := []struct {
-		name               string
-		selectorValueRegex string
-		image              string
-		assertion          func(t *testing.T, found bool)
-	}{
-		{
-			name:               "match any image",
-			selectorValueRegex: ".*",
-			image:              "alpine",
-			assertion:          assertShouldMatch,
-		},
-		{
-			name:               "match image with same tag",
-			selectorValueRegex: ".*:latest",
-			image:              "alpine:latest",
-			assertion:          assertShouldMatch,
-		},
-		{
-			name:               "don't match image with different tag",
-			selectorValueRegex: ".*:latest",
-			image:              "debian:jessie",
-			assertion:          assertShouldNotMatch,
-		},
-		{
-			name:               "match image by exact name",
-			selectorValueRegex: "alpine",
-			image:              "alpine",
-			assertion:          assertShouldMatch,
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			found := doesMatchImageResource(testCase.selectorValueRegex, testCase.image)
-			testCase.assertion(t, found)
-		})
-	}
-}
-
-func testMetadata() metav1.ObjectMeta {
-	labels := map[string]string{
-		"labelkey":   "lvalue",
-		"labelkey2":  "lvalue2",
-		"labelowner": "lsometeam",
-	}
-
-	annotations := map[string]string{
-		"annotationkey":   "avalue",
-		"annotationkey2":  "avalue2",
-		"annotationowner": "asometeam",
-	}
-
-	return metav1.ObjectMeta{Labels: labels, Annotations: annotations}
-}
-
-func assertShouldMatch(t *testing.T, doesMatch bool) {
-	t.Helper()
-
-	if !doesMatch {
-		t.Fatal("Failed to match")
-	}
-
-	t.Log("Matched all properly")
-}
-
-func assertShouldNotMatch(t *testing.T, doesMatch bool) {
-	t.Helper()
-
-	if doesMatch {
-		t.Fatal("Incorrectly matched")
-	}
-
-	t.Log("Correctly did not match")
-}
-
 func TestValidate(t *testing.T) {
 	testCases := []struct {
 		name                      string
